requirements.txt
Django>=3.1.5,<3.2
djangorestframework>=3.12.2,<3.13
psycopg2>=2.8.6,<2.9
requests>=2.25.1,<2.26